cloneprod: Apply the audit schema when cloning prod.
Since the audit tables are owned by a separate database user,
they do not get cloned on exp/imp.  This applies the audit schema
update after cloning.

An alternate would be to export the audit schema and then re-import
it.  For now, this is easier.  Just needed a quick tweak to the old
upgrade script so that sqlplus will finish cleanly.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>