chooser: Validate the machine plenary.
Addresses-Issue: Jira/AQUILON-477

Whenever the Chooser is called to verify service bindings on a host
(including aq make and reconfigure, but not compile), the machine
plenary will be included as one to be verified.

This is something of a hack in lieu of a more generic mechanism to
continually validate, rewrite, and recompile templates as needed.
It is meant to provide a back-up mechanism for any commands that
may not be correctly re-writing the template as needed.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>