authorization: Allow operations to add and delete non-internal networks.
This removes the restriction that a user must be an aqd_admin to add
or del a network, but introduces a config option to restrict updates
to the default (internal) network environment to certain roles.

I don't think entitlements should look like this... it's just a hack
until we have something better.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>