Allow binding the broker to a specific IP address
When bind_address is set in the configuration, then the broker will only
listen on the specified address, and notifications will also be sent out
from this address.

Sending out notifications to a large number of hosts will now use the
same socket instead of closing/reopening it all the time.

Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>