add_machine: Add a new prefix option.
There is a need for an atomic operation that combines 'search_next'
and 'add_machine' in order to make the operation atomic. This commit
adds a new prefix option that works the way the machine option does
in search next.  In this way, it can get the next machine name and
reserve it in one transaction.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>