runtests: Fix use of --mirror when using default config.
I've been meaning to fix this for ages.  No need to pollute the
environment of the spawned caller - either the script was called
with an alternate config (which will be passed with the rest of
the args), or it was called with the default (which will work
just fine after the mirror).

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>