reconfigure: Fix for personality change of clustered hosts.
The personality being tested was wrong (need to test that the new
personality is allowed).  The final test for whether a host is taking
on an allowed personality for the cluster had a typo.  The error message
that the test is supposed to generate was also missing data.

Similar problems existed in reconfigure_list.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>