show_host: Include server binding information.
I thought there was a jira for this, but apparently not.

Include any service instances that the server provides in the raw output.
The output now includes 'Provides: ' lines for the service instances
that the host is bound as a server for.

I think it would have been better just to have extra Template lines,
but this would have required fixing the client paths to include
/client/config and I'm worried about breaking scripts.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>