authorization: New hpevelo role.
Added a role to help manage the HPE velocity lab.  The name is
horrible but this is all something of a hack.

The proid using the role needs to be able to reconfigure hosts
into different personalities and OS versions.  Due to a problem
with the lab hardware presenting disks as cciss on RHEL5 and
scsi on RHEL6 the role also needs to be able to add/del disk.

Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>