Fixes for missed sequences from past upgrades.
The 1.8.0 upgrade neglected to add two required sequences for the
resource and resholder tables.
The 1.6.5 upgrade neglected to remove a sequence for vlan_interface
when the table was dropped.

The adds were done prior to the 1.8.5 upgrade, and the remove was
done during the 1.8.5 upgrade (since I had it listed in the TCM),
so these are already in prod.  Instead of putting the upgrade
scripts in a this_branch directory I just put them directly in
the 1.8.5 tree.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>