Add resourcegroup resources to AQDB
Description: In order to model vcs clusters properly in AQDB we need
to be able the concept of "service groups" which are a collection of
resources which work together to provide a service. Since the term
service is already used to mean something else, and to avoid vcs-specific
terminology, we will call these "resourcegroups". Since resourcegroups
are both a resource and a resource holder, the corresponding resourceholder
is called BundleResource rather than ResourceGroupResource.

Also added a new unit test to exercise the new options.

Test Plan: Ran unit tests and added resources in a dev broker.

Addresses-Issue: AQUILON-508
Code Review: http://vorlon.ms.com:8000/r/140/

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>