Add new status for host to be used by external components (i.e ESP)
Addresses-Issues : JIRA/AQUILON-437

This fix adds a new boolean flag (advertise_status) to hosts. The flag by default
is false build and is marked to true once the host is in ready state.
This state advertised in the dtawarehouse code will be 'build' if the flag is false
and whatever the object status is if the flag is true.
The state can be reset to flag by a admin when the host needs
to be rebuilt etc and reflect as such in datawarehouse/ESP.
As such the only api method available for maintaining this status is
reset_advertised_status.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>