Let --configure be the default for pxeswitch
Addresses-Issue: Jira/AQUILON-561

An out of date kickstart file can lead to hard to detect build problems,
therefore it makes sense to make --configure the default behavior for
the pxeswitch command. The old behavior can be restored by using the
--noconfigure flag.

As a consequence, specifying an action for pxeswitch is now optional,
and the --status flag automatically turns off --configure.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>