Correctly validate resgroup may not be held by another resgroup
@validates methods must be set in the class that defines the column so
the holder attribute must be validated in the Resource class. However
at the point the method is called there is no way to determine the type
of the resource because the resource_type attribute is not set until the
session is committed().

The workaround is to instead perform the validation on the resource_type
field. This means the error won't be caught until commit() rather than
at the point when holder is assigned to.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>