Do not write plenaries for non-compilable hosts
Addresses-Issue: Jira/AQUILONAQD-405

There are some corner cases in update_machine when a host plenary will
be written even if the archetype is not compilable, which will cause
compiling the domain to fail. Adding the logic to avoid writing the
plenary in such cases to the template code means we don't have to worry
about the individual commands.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>