Change schema upgrade scripts to be Oracle dialect
The previous attempt at this was using the sqlite dialect. This uses
oracle dialect and has been tested on the oracle test database.

I also noticed as part of this that I had misnamed one of the
constraints in resourcegroup.py, re-using the name from the filesystem
code. Oracle objected to that and so I renamed it.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>