add_intervention: Have resource template include name.
This makes interventions more generically useful.  The specific case
I have in mind is zephyr VMs - zephyr could add a zephyr_lifetime
intervention with an expiry that matches the reclamation time.  The
templates just need to know how to deal with an intervention with
the right name.  In this case, it's just a change of the motd. :)

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>