update_interface: Remove ability to add a new IP.
Addresses-Issue: Jira/AQUILONAQD-344

It no longer makes sense for update_interface to be able to add an
IP address to an interface without context.  Using the command this
way ends up creating odd constructs in the database and has left a
mess to be cleaned up in the Velocity lab.  Applying this patch
suggestion from Gabor now to stop the bleeding.