Split out the DNS entry allocation logic
There are multiple places in the code which either want to use an
existing DNS record, or add a new one if it's missing. This task is
surprisingly complicated if we want to handle all the corner cases, so
it makes sense to add a common helper function. This commit moves the
code from add_interface_address to the common helper; the logic is not
changed.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>