Adjustments for tests that use the local host name.
I'm hesitant to call them fixes. :)  The use of socket.getfqdn()
was a nice idea, but python doesn't try verify hard to actually do
what the method name suggests.  We could write that logic ourselves,
but since this is for a unit test it seems sane to just allow a
config file override for any cases that need it.

Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>