config: Switch hostname default to getfqdn.
The config was using socket.gethostname as a default for the hostname.
This works on aquilon nodes where the hostname comes back as fully
qualified by default.  On aurora it causes some pain for the regression
tests.

Interestingly enough, it only shows a difference on the aurora RHEL 5
builds (not RHEL 4).  To test on RHEL 4 a manual config override is
still needed.  However, that makes this a safe change for production,
as aqd is currently running on an aurora RHEL 4 VCS cluster.

Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>