Clean up DB connection setup
The password file is now always read if it is set. If the password file
is not set, the password defaults to be empty. Drop support for
specifying the password directly in the config file.

Standardize on using the SQLAlchemy dialect name when checking for
DB-specific behavior if there's already a DB connection, and checking
the beginning of the DSN if there's no DB connection.

Change the default Oracle configuration to assume Kerberos.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>