Use the inet type for storing IP addresses on Postgres
SQLAlchemy's Integer type maps to a 32-bit DB type on Postgres by
default, which causes problems when Python is built for 64-bit and
therefore int() values may be too large. PostgreSQL has a native 'inet'
data type, so use that instead of Integer. The max() function however
does not work for inet, so use order() and limit() to implement max().
The query cost does not change significantly on Oracle, so this should
be safe.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>