search_next: Fix the pack algorithm for similar names.
The --pack logic expects to iterate over a list like (1, 2, 3, 5)
and "fill in the hole" with (in this example) a 4.  However, the
sql query to pull names might result in names like host1, host1-e1,
host1-e2, host2, and thus a list like (1, 1, 1, 2).  The extra
entries broke the logic.

The fix is to not store the duplicates.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>