search_next: Restrict candidate match list.
The idea behind search next is that it searchs for the given prefix
followed by a number.  However, the regex implementation allowed for
non-numeric characters to sit between the prefix and the number.
This would allow for foobar1 to count when searching for foo.  The
fix is to anchor the regex at the beginning of the input.

The regex still allows for any characters to follow the number, as
I think we want to "count" an entry like foobar1-e1, even if foobar1
doesn't actually exist.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>