python_patches: Speed fix for loading the uuid module.
Introduce a new module to store monkey-patch style solutions
for problems with the stock python libraries.  In this case,
the use of ctypes.util.find_library on import of uuid causes
a notable increase in start times.  A helper is provided to
temporarily monkey-patch the find_library method while
importing the uuid module.

Because the aq client loads this module, there is a savings
of .07 seconds every time aq is run, adding up to about
five fewer minutes running unit tests.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>