Rework the running of DSDB commands
Instead of running the commands immediately, the new interface allows
the commands to run to be queued up and then run all at once. Every
command is now responsible to register its own rollback action as well.
This means all the rollback logic is now contained in the DSDBRunner
class.

Reviewed-By: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>