Clean up resource plenaries
Calling Plenary.get_plenary() for the same DB object twice currently
returns two different instances of the Plenary class. If .remove() is
called on the first instance and then .write() is called on the second,
as happens with resource groups, then the plenary template will be
written out even though it should be removed. Changing the order fixes
it for now.