Feature handling in personality
Description:
- Originally features were mapped at personality but were generated as part of host template.  Now the features
are mapped at archetype or personality and will be included as part of plenary personality generation.
The show functionality remains the same where the host shows the features mapped to it
thru archetype or personality, but the tests which check host plenaries have been removed.
Simillarly the flush option for bind feature has been removed the personality plenary
will be immediately generated.

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>