tests: Remove last rsync to nyaqd1.
The tests need some templates that are now available over git from
the prod broker.  This change removes the last requirement on the
tests for login access to the broker.

Anyone with a custom config file for tests (such as for using
Oracle) will need to add the new value in unittest.conf for
unittest/swrep_repository.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>