refresh_windows_hosts: Added missing continue.
There was an error handling block that was clearly meant to move
on to the next host entry, but the continue statement was either
missed or refactored.

We saw this triggered in prod when adding an msad.ms.com alias
where the target does not exist.  I replicated this scenario as a
test.  One item of note is that I marked the msad.ms.com domain
as restricted to mirror prod and to get the behavior of creating
a ReservedName.  This had no effect on any other tests.

Addresses-Issue: Jira/AQUILON-794

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>