tests: Add a reconfigure test.
The change made in 4086c44f was being verified with a compile of
a host where the machine plenary should have changed.  While
testing the bump to twisted 12.0, it was noticed in one environment
that this test failed.

The new test doesn't change anything, it just verifies that the old
state of the plenary is different from another existing test for the
new state.  The hope is that this will help track down the issue.

Unfortunately, it may instead have the effect of masking the problem.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>