Add SQL upgrade script for node_index
Add an SQL upgrade script which adds the node_index column to the
host_cluster_member table and populates it for existing clusters.

Tested by running the SQL manually against the NY Oracle test database.

Also correct docstring for HostClusterMember class to make it clear this
is an asssociation table not proxy.

Addresses-Issue: Jira/AQUILON-634

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>