Change parameter formatting
Remove misleading header lines - we're displaying the paremeters, not
the archetype or feature link. Anyway, FeatureLink is an internal data
type users should not know about.

Stop calling the formatter "json", because it is not.

Drop unneeded ParameterList class, the default list formatter is good
enough.

Remove extra space before ':'.

Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>