Implement switch discovery
Switch discovery is implemented by calling an external helper script
that gathers the information, and then produces a JSON output the broker
consumes. This way the differences between switch vendors and models are
hidden from the broker.

The "show_switch --discover" command outputs the list of aq commands the
user should invoke to make the information in AQDB match the discovered
state of the switch. The "update_switch --discover" command performs all
the updates in one go.

Change-Id: I6f490b93b3e89fda554cd58d12cb7a974e339f80

Reviewed-By: Nathan.Dimmock@morganstanley.com