Relax name check of loopback interfaces
There does not seem to be a common naming convention for switch loopback
interfaces, so remove the restriction that the name of such interface
should start with "loop". Keep the logic in add_switch and add_interface
to default the interface type to loopback if the name starts with "lo".

Change-Id: Ic62941afc4aaaa72b003fe42a1f1e50b5a8f46d9

Reviewed-By: Nathan.Dimmock@morganstanley.com