Do not try to optimize static routes
The broker previously omitted static route definitions if the gateway of
the route was also the default gateway. However if there's a static
route covering a large block of networks pointing to a non-default
gateway, and there are smaller routes pointing to the default gateway,
then omitting those routes is wrong.

Addresses-Issue: Jira/AQUILON-853

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>