Expand mssb-unixops role to accommodate IP reservation.
The current mssb-unixops role needs to be expanded to allow for
IP reservation.  The IP reservation will be done using add_address.
Once the inventory information such as (mac address, location etc) is
available for the host the address will then be converted to a proper
host.  This requires the del_address and del_alias commands.

Addresses-Issue: Jira/AQUILON-868

Reviewed-By: Wesley Hertlein <Wesley.Hertlein@morganstanley.com>