runtests: Add a check for configured hostname.
In commit 5d27c4da several tests started relying on the hostname
as detected by the config defaults.  This broke tests on our Aurora
RHEL 4 systems where a shortname gets detected but the tests
are expecting a fully qualified name.

I knew this at the time.  I might even have documented it
somewhere, but this error still caught me when I hit it today.

This patch checks the value up front and provides guidance on
a workaround if a shortname is detected.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>