Add a new environment field on personality
The new environment field would be used to determine the
host environment within the personality. This field should not
be used to determine any for of host sla, which would be tracked/managed
seperately. The env field besides including the tam based values for
host will included a value of legacy which would be the default value
for all existing personalities. Once when we are at the point that all
personalities have the appropriate values we probably should drop it.
This also includes cleanup of unused personalities from unittest.dmp

Addresses-Issue: Jira/AQUILON-886

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>