Reordered the tests and added few more tests
The tests were reordered such that host profile will fail if
a required parameter in the schema is not set. The host profile
compiles once the required parameter is set. There are other cases where
parameters are required for other downstream processes but may not
be required for the compilation, but these test do not test those cases.

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>