New role for spot_server
The new tool spot needs to be able to reserve an ip a day prior
and then create a host using that reservation and then
destroy the host later just like zephyr. This role adds
the command for this. Speaking to Josh I have also added
some of the additional commands zephyr needs such that we could
at some move the zephyr prodids under this role.

Addresses-Issue: Jira/AQUILON-909

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>