Fix for Command: del archetype, dependencies
The dependencies for archetype are now queried for; and in case found, the
brocker will issue a message saying what the dependencies and their
names are.  Previously an error message, with the delete constraint
violation, was displayed.

Added test cases to add_os and add os to archetype so that archetypes with
cascading dependencies will be deleted.  Added delete cascade for operating
system.

Addresses-Issue: Jira/AQUILONAQD-400

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>