Clean up input.xml for add_cluster
The usage information output when you fail to specify all the mandatory
options for add_cluster is not as clear as when you get the same error
from add_host. This diff makes add_cluster consistent with add_host and
also re-orders it to make it easier to follow.

Addresses-Issue: Jira/AQUILON-434

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>