cluster state change to decommissioned is propagated to vm hosts.
Changing the status of a cluster to decommissioned should also set the statuses of the vmhosts.
"aq change status" should refuse to set the status to "decommissioned" if the host or cluster has VMs.
A special case: a vmhost could be set to decommissioned state if it or it's cluster has no VMs.

Addresses-Issue: AQUILON-906

Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>