Fix port group allocation
The previous pg allocation algorithm counted VMs on a per observed VLAN
basis. However, if the same network/VLAN was observed on multiple
switches, but only one of the switches is actually bound to an ESX
cluster, then the VM count for that observed VLAN will always be 0, so
the algorithm always preferred this VLAN thinking it's not used. That
broke the round-robin allocation of VLANs.

When checking which VLAN is the least used, we need to do that on a
per network basis, summing allocations registered to any other switches.

Addresses-Issue: Jira/AQUILON-684
Change-Id: I042e89441f44a1490ee37c768428653cedd5c617
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>