Make the template extension configurable
Quattor upstream prefers the '.pan' extension over '.tpl' for some time
now. Allow the extension to be set in the configuration file. There are
some pieces of code (mostly in the unit test) dealing with the
hand-written templates where the broker has no control over the
extension, such code has to allow both '.tpl' and '.pan' for the
transition period.

Change-Id: I0e78062b3fe1c8aad015c385058b80242b5777ae
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>