Allow deferring a resourcegroup constraint
There's a circular dependency between the resource and resholder
tables when there are resourcegroups present. We have to break this
dependency when data is copied from one DB to another using the
aqdb_migrate script. aqdb_migrate should also do the full copy inside a
single transaction, because COMMIT would activate the constraint.

Change-Id: I7dad9e5d83d180a2dfde63066fc89d38e503bc0c
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>