Fix ARecord.assignments
The previous construct worked with 0.7, but does not generate the
correct query with 0.8 (thanks to the ORM the result is correct, but it
is slow). Since nothing in the documentation suggests the original form
should work, it's better to change it to something supported.
Fortunately, we can just make the query simpler, as in the
ARecord->AddressAssignment direction we don't have to worry about DNS
environments - whether a given IP address is in use or not does not
depend on DNS visibility.

Change-Id: I93ccd2dfc27be9312d5dc4e3a9ce7abfa5bd1905