Fix tests broken due to personality template fixes.
The tests use actual personality templates while testing.
with parameters for personality moving to aqdb, the data for
personality need to be set up in db while testing. As a result
the param definitionis have moved out of parameters and in a seperate
setup file. This has also shuffled and changed some of the
parameter tests.
Also the errors due to missing templates have
changed and instead of erroring out with template not found
we get errors for missing personality parameter.

Change-Id: If66b57d9d947f05f20211f3b5ac463a9dce62ed4
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>