Set the primary key name by default
Use __declare_last__() to initialize the name of the primary key if it
is not set explicitly. This allows removing quite a bit of code, and
reduces the need for pylint overrides.

There are a couple of cases when the default algorithm would generate a
name that's too long, leave the existing override in place for those.

Change-Id: I0446352785f9fd999adf2068a17991dfedbeffcc
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>