Add hack to handle deleted objects
As we make the calculation of the template name more dynamic, we must
protect from the case when there are multiple instances of the same
Plenary subclass for the same DB object, one is notified about the
deletion of the object, but the other is not. This is really just a hack
to supress the errors.

Change-Id: Id24b49d5be89597a29e503e7f0bf1677645c378e
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>