Drop Plenary.cleanup()
The cleanup() method was used to remove build files, using the old
domain/sandbox location if the host or cluster was being moved into a
new domain/sandbox. With the new dynamic path calculation in place, we
can just create the Plenary objects early and let them remember the
original location.

Change-Id: I094552a98c5ddef48c93d4f15a1082ff10532f6e
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>