Finalize the host environment
Fully db personalites with 'legacy' host env print
out host env path in template, whereas the non-db
personalites do not have this path in template. To
be consistent we shouldnt print out the path for
personalities marked with host_env 'legacy'.
Host env should not be mutable so mark it with
final.

Addresses-Issues: Jira/AQUILON-1443
Change-Id: I159fb939a664dbe64f6404941c170c055bf09e74
Reviewed-By: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>