Move protobuf definitions to input.xml
This makes it possible to discover the expected message format on the
client side.

The loading of the protocol modules moved from ObjectFormatter to
ResponseFormatter. The signature of the format_proto() method is
changed, the message container is now always passed in, and nothing is
returned as the structure is modified in-place. As a side effect, the
standard ListFormatter now works just fine, we don't need to override
format_proto() in subclasses of ListFormatter anymore.

Change-Id: Id05245c7982db581d5f9586821078f74888d6c28
Reviewed-by: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>