Enforce interface name uniqueness on add_interface_xxx comamnds
Each of the add interface commands should guard against attempts to add
an interface that already exists.  This is done in many other places in
the code using get_unique(preclude=True).  I have followed suit for
consistancies sake.  I have modified add_interface_machine to have the
same behaviour, as it was doing its own thing.

The code in get_or_create_interface is correct.  However the columb type
of interface.name is AqStr; this forces everything to lower case.  The
comparisons assume that supplied (interface) name is the same format;
yet this is not enforced (bug for another day).  As a fix I have changed
the supplied name to be lower case.

Change-Id: I1ef15a7c8b8d0abf913dada28bd984656a786694
Addresses-Issue: Jira/AQUILON-666
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>