Cache plenary objects
Having multiple, different plenary objects for the same plenary template
file is problematic, and caused issues in the past. There are some use
cases (e.g. touching a resource plenary that is also part of a resource
group) when it is difficult to control the instantiation of plenary
objects, so build an internal cache instead.

Change-Id: If8eff5a05ad21e7f3695a5f60f6a4e20b0b426cf
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>