Remove the sandbox warning
Currently the sandbox checks against the default prod domain
to check if it has the latest commit and then prints a
message to merge against prod domain. Unfortunately this
is not true as we have bunch of domains and the message is incorrect
and confusing in lot of the cases. We should drop this
message or alternatively store the start domain and do the checks
against the domain the branch was started from. Currently
we only retain a git commit id and its impossible to
determine which domain or sanddox the branch was started from.

Addresses-Issue: AQUILON-1508
Change-Id: I9b2b8523091db97b93f0bf95318f23dfeaf10950
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>