Fix add_sandbox checking and erorr handeling.
When creating a sandbox its possible to leave the broker in an incosistant
state.  This occurs when CommandGet (executed at the end of add_sandbox)
fails as the sandbox dir already exists.  In this case the database mutations
(addition of sandbox) is rolled back; however, the branch has been added to
git.  The result is that branch remains created in the template-king yet the
database does has no knowlege of its existance.

This change adds an additional check to add_sandbox to ensure that the
target directory does not exist before starting.  In addition the database
state is commited after the git command succeeds.

Addresses-Issue: Jira/AQUILON-925
Change-Id: I049ecb4b82a0bbe2ec88e51be0d4fc41755bbf8e
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>