Consistency Checking
There have been a number of times when I have thought we needed a consistency
checker; this is a starter for ten!

This commit adds a consistency checker that checks branches (domain and
sandbox) in git, the database and filing system.  In theory they should
all be consistent; they probably aren't.

Change-Id: I931ed022d9b391f08707a7cbd635867d74b373c0
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>