Latency threshold attribute added to Share.
A new attribute called latency_threshold is added to Share
which is a threshold in millesecs.
If this threshold is exceeded, vSphere would automatically
throttle the VM consuming I/O bandwidth.
This commit effectively reverts 307503c.
Latency threshold will default to 20 and should be a non-zero value.
Latency threshold needs to be updated from time to time so we need a
command to update share attributes.

Addresses-Issue: Jira/AQUILON-1539
Change-Id: Ie370902c904b10798eef73c35b271dab87240db8
Reviewed-By: Zoltan Sebestyen <Zoltan.Sebestyen@morganstanley.com>