Fix DSDB comment propergation
This changes the way that 'dsdb update_host' is called.  If the
update includes an IP address of a MAC address then the interface
name is passed during the update (if supplied).

When only the comments change, then no interface name is supplied,
thus DSDB will update the host comment and not the interface comment.

Addresses-Issue: Jira/AQUILON-1382
Change-Id: I450de19bf853f16218eb998cfd89b6a3e9a23ee1
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>