deprecated warnings combined with --noauth flag results in Internal Server Error
There's a logic in the broker to log the name of the user specifying a
deprecated command or option. In case of the readonly broker username is
not set, resulting in a 'Too few arguments' error.
Currently, for performance reasons, these deprecation messages are
disabled for show commands.

Change-Id: I8a6ed2fed6689e9bda47a1d476324b458b0b359d
Addresses-Issue: Jira/AQUILON-1592
Reviewed-By: Priya Joshi <Priya.Joshi@morganstanley.com>