Improve handling of single-instance classes
Allow the get_instace() method to be called on the base class, taking
the name of the subclass to look up. Replace the use of get_unique()
with get_instance() where possible, because get_instance() is cheaper,
and it also handles normalization (e.g. mixed-case input) better due to
the internal use of polymorphic_subclass().

Change-Id: If687ad5b329e41887d964a9b34f81493d7954eea
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>