Move parameter checks earlier
Perform the base syntax check and type conversion before we call the
render method. This allows dropping BrokerCommand.parameters_by_type and
BrokerCommand.parameter_types. As a consequence, less processing is
needed when inserting records in the audit tables. As a side effect,
malformed options will not be added to the audit table.

Change-Id: Ib6116f5e7599940462289d92f464bc6ffce3e799
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>