map_grn: Avoid throwing an internal error
Return an error message if someone tries to map a GRN to an archetype we
don't support, instead of throwing an internal error.

Update host plenaries in addition to personalities, since the host
plenaries also have a copy of the GRNs. This needs changes in the
testsuite as well, as map_grn will now give a warning if a host does not
have all the service bindings set up.

Change-Id: I3abeed92a6c70189980542990adf1a903c165ad4
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>