Reduce noise from failed compilations
Messages from the compiler are printed while the compilation is still
running, no need for duplicating them if the compilation fails.

There's an edge case if the compiler cannot even start, but that should
be rare enough that we don't have to worry about it here - one can
always look at the broker's logs for the full story.

Change-Id: Ia5c474b7217e069a6c70262a8a2028b30ac97003
Reviewed-by: Priya Joshi <Priya.Joshi@morganstanley.com>
Reviewed-by: Dave Reeve <Dave.Reeve@morganstanley.com>