Refactor 'host' formatting
This change moves the raw formatter for a host object into a new
method in the hardware entity class.  The current host formatter just
delegates to the associated hardware entity, which may or may not
include a host.  This hardware entity formatter can now call this
common method to display the host information.

Change-Id: I38e6dbbb84d35ff6f8522e568d69ab0259eb90c4
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>