Reloacte input.xml parsing in seperate class
Before this change the parsing of input.xml was done inside
the __init__ method of RestServer.  This change moves this
parsing into a CommandRegistry class which calles insert_instance
on the RestServer to record new commands.

The intention is to refactor this such that CommandRegistry
provides a base class for reading input.xml.  Through the use
of callbacks and helper classes this will allow different
command registries to be populated.

Change-Id: I6ad786c6dfbe9e7e148572db08f1c78e04d1d7f8
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>