Extend CommandRegistry to handle argument parsing and processing
This change intorduces the concept of a CommandEntry; which is created for
each command transport in input.xml by the CommandRegistry.

The CommandRegistry uses its methods new_entry and add_entry to create new
CommandEntry's and add them to the registry (RestServer in this case).  The
intension is that this can later be abstracted to allow for other storage
classes.

The CommandRegistry invokes add_option and add_format for each option and
format it finds for a given command.  This moves the processing logic from
the CommandRegistry into the CommandEntry.  Longer term the intention is to
abstract the CommandEntry into a base class.

The check_arguments function from RestServer has also been incorporated, as
it how has all the information it requires in the CommandEntry.  This required
a small change to the request handeling.

Change-Id: Idfbb76503e171b9af60bbd7ffe3df0ad33253bc5
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>