Restore interface name/mac on DSDB rollback
If the DSDB deletion of a network device or chassic failes then
the interface and mac address information will be lost.  This
change modifies the behaviour of DSDBRunner.update_host to
handle deletion (by passing only the old information).  The
snaphot created with DSDBRunner.snapshot_hw contains all of the
information needd to rollback; removing the need to local variables
to contain the same information.

Change-Id: I36479ce4069a463735fb2ea5b413610605bef499
Reviewed-by: Gabor Gombas <Gabor.Gombas@morganstanley.com>