Lookup request status directly in show_request
At the moment show_request displays the status of a request by retrieveing
the status object attached to its logger.  This works as show_request
is handled differently, in that its status object is in fact the same object
as the request its monitoring.  This change causes show_request to lookup the
status object directly in the cataloge.

Change-Id: Ic8d22405cc7cdc756f844a54fc611255060b6873
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>