Creating status cataloge entries for show_request
This commit creates as seperate entry in the request catalog for show
request commands.  Previously a show request command would share the
entry with the comamnd it was viewing.

The logger created for the show_request command now logs its output
(of which there is none unless there is an error) to its own status
catalog.  This avoids the confustion of cross logging with the
monitored command.

As a result of the new entry the special handeling logic from disposing
of the logger is not longer required.  Previouly only the removal of
the entry by auditid (called by the monitored comamnd) would remove
the entry in the status catalog (after a short delay).  Where as removing
the entry by the requestid (called by the show_request command) would
only dispose of the request id lookup in the catalog.

Now the status catalog entry is only removed by the command itself.

Finally the show_active_commands command has been modified to only
display commands that have not 'finished'; previously it relied on
the above logic to correctly display its output.

Change-Id: I7256b89d3772dc8ca61622d389bfef6477715ae2
Reviewed-By: Gabor Gombas <Gabor.Gombas@morganstanley.com>